+++ apps/chaingraph-backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @badaitech/chaingraph-backend
 
+## 0.1.5
+
+### Patch Changes
+
+- Fix and adjust chaingraph-frontend lib NPM package
+- Updated dependencies
+  - @badaitech/chaingraph-nodes@0.1.5
+  - @badaitech/chaingraph-trpc@0.1.5
+  - @badaitech/chaingraph-types@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes